LF-5217 Move all logic into new FarmNotes container
