+++ packages/webapp/src/containers/FarmNotes/index.tsx
@@ -1,0 +1,121 @@
+/*
+ *  Copyright 2026 LiteFarm.org
+ *  This file is part of LiteFarm.
+ *
+ *  LiteFarm is free software: you can redistribute it and/or modify
+ *  it under the terms of the GNU General Public License as published by
+ *  the Free Software Foundation, either version 3 of the License, or
+ *  (at your option) any later version.
+ *
+ *  LiteFarm is distributed in the hope that it will be useful,
+ *  but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ *  GNU General Public License for more details, see <https://www.gnu.org/licenses/>.
+ */
+
+import { useState } from 'react';
+import { useSelector } from 'react-redux';
+import { useTranslation } from 'react-i18next';
+import { userFarmSelector, userDisplayNameMapSelector } from '../userFarmSlice';
+import { useDeleteFarmNoteMutation, useGetFarmNotesQuery } from '../../store/api/farmNoteApi';
+import {
+  useGetFarmNotesReadQuery,
+  useMarkFarmNotesReadMutation,
+} from '../../store/api/farmNotesReadApi';
+import type { FarmNote } from '../../store/api/types';
+import FarmNoteFormContainer from './FarmNoteForm';
+import FarmNoteList from '../../components/FarmNotes/FarmNoteList/';
+import FarmNotesFloatingButton from '../../components/FarmNotes/FarmNotesFloatingButton/';
+import Drawer, { DesktopDrawerVariants } from '../../components/Drawer';
+import ImageLightbox from '../../components/ImageLightbox';
+import styles from './styles.module.scss';
+
+type FormState = null | 'add' | { mode: 'edit'; note: FarmNote };
+
+export default function FarmNotes() {
+  const { t } = useTranslation();
+  const userFarm = useSelector(userFarmSelector) as { user_id?: string; [key: string]: unknown };
+  const userDisplayNameMap = useSelector(userDisplayNameMapSelector);
+
+  const { data: farmNotes } = useGetFarmNotesQuery();
+  const [deleteFarmNote] = useDeleteFarmNoteMutation();
+
+  const { data: farmNotesRead } = useGetFarmNotesReadQuery();
+  const [markFarmNotesRead] = useMarkFarmNotesReadMutation();
+
+  const [isDrawerOpen, setIsDrawerOpen] = useState(false);
+  const [formState, setFormState] = useState<FormState>(null);
+  const [lightboxSrc, setLightboxSrc] = useState<string | null>(null);
+
+  const handleOpenDrawer = () => {
+    setIsDrawerOpen(true);
+    markFarmNotesRead();
+  };
+
+  const handleCloseDrawer = () => {
+    setIsDrawerOpen(false);
+    setFormState(null);
+  };
+
+  const handleDeleteFarmNote = async (note: FarmNote) => {
+    try {
+      await deleteFarmNote(note.id).unwrap();
+    } catch (error) {
+      console.error('Failed to delete farm note:', error);
+    }
+  };
+
+  const isFormOpen = formState !== null;
+  const isEditState = formState !== null && formState !== 'add';
+  const formTitle = isEditState ? t('FARM_NOTE.EDIT_NOTE') : t('FARM_NOTE.ADD_A_NOTE');
+
+  const lastReadAt = farmNotesRead?.last_read_at;
+  const hasUnread = farmNotes?.some(
+    (note) =>
+      note.user_id !== userFarm?.user_id &&
+      (!lastReadAt || new Date(note.updated_at) > new Date(lastReadAt)),
+  );
+
+  return (
+    <>
+      <div className={styles.fab}>
+        <FarmNotesFloatingButton onClick={handleOpenDrawer} hasUnread={!!hasUnread} />
+      </div>
+
+      <Drawer
+        isOpen={isDrawerOpen}
+        onClose={handleCloseDrawer}
+        title={isFormOpen ? formTitle : t('FARM_NOTE.FARM_NOTES')}
+        desktopVariant={DesktopDrawerVariants.SIDE_DRAWER}
+        fullHeight
+        addBackdrop={false}
+        classes={{ desktopSideDrawerContainer: styles.sideDrawerContainer }}
+      >
+        {isFormOpen ? (
+          <FarmNoteFormContainer
+            key={isEditState ? formState.note.id : 'add'}
+            note={isEditState ? formState.note : undefined}
+            onSuccess={() => setFormState(null)}
+            onCancel={() => setFormState(null)}
+          />
+        ) : (
+          <FarmNoteList
+            notes={farmNotes || []}
+            userDisplayNameMap={userDisplayNameMap}
+            currentUserId={userFarm?.user_id ?? ''}
+            onAddNote={() => setFormState('add')}
+            onEditNote={(note) => setFormState({ mode: 'edit', note })}
+            onDeleteNote={handleDeleteFarmNote}
+            onImageClick={(src) => setLightboxSrc(src)}
+          />
+        )}
+      </Drawer>
+
+      <ImageLightbox
+        src={lightboxSrc || ''}
+        open={!!lightboxSrc}
+        onClose={() => setLightboxSrc(null)}
+      />
+    </>
+  );
+}
