--- packages/webapp/src/containers/Home/styles.module.scss
@@ -1,11 +1,0 @@
-.fab {
-  position: fixed;
-  bottom: 24px;
-  right: 24px;
-  z-index: 10;
-}
-
-.sideDrawerContainer {
-  width: 520px;
-  max-width: 100%;
-}
